% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour.with.holes.R
\name{create_habitat_with_holes}
\alias{create_habitat_with_holes}
\title{Create a habitat with holes}
\usage{
create_habitat_with_holes(outer, holes)
}
\arguments{
\item{outer}{A two-column matrix of coordinates (longitude, latitude).}

\item{holes}{List of holes. Each hole is a two-column matrix of coordinates
(longitude, latitude).}
}
\description{
Create a habitat with holes
}
