% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rEEMSplots.R
\name{eems.plots}
\alias{eems.plots}
\title{A function to plot EEMS results}
\usage{
eems.plots(mcmcpath, plotpath, longlat, plot.width = 7, plot.height = 7,
  out.png = TRUE, res = 600, xpd = TRUE, add.grid = FALSE,
  col.grid = "gray80", lwd.grid = 1, add.demes = FALSE,
  col.demes = "black", pch.demes = 19, min.cex.demes = 1,
  max.cex.demes = 3, add.outline = FALSE, col.outline = "white",
  lwd.outline = 2, projection.in = NULL, projection.out = NULL,
  add.map = FALSE, col.map = "gray60", lwd.map = 2, eems.colors = NULL)
}
\arguments{
\item{mcmcpath}{A list of EEMS output directories, for the same dataset.}

\item{plotpath}{The name of the graphics files to generate. There are five output figures: plotpath-mrates01 (effective migration surface), plotpath-qrates01 (effective diversity surface), plotpath-rdist01 (between-demes component of genetic dissimilarity), plotpath-rdist02 (within-demes component of genetic dissimilarity), plotpath-pilogl01 (posterior probability trace).}

\item{longlat}{A logical value indicating whether the coordinates are given as pairs (longitude, latitude) or (latitude, longitude).}

\item{plot.width,plot.height}{The width and height of the graphics region for the two rate contour plots, in inches. The default values are both 7.}

\item{out.png}{A logical value indicating whether to generate png files or pdf graphics files.}

\item{res}{Resolution, in dots per inch; used only if out.png is set to TRUE. The default is 600.}

\item{xpd}{A logical value indicating whether to clip plotting to the figure region (xpd = TRUE, which is the default) or clip plotting to the plot region (xpd = FALSE).}

\item{add.grid}{A logical value indicating whether to add the population grid or not.}

\item{col.grid}{The color of the population grid. Defaults to 'gray80'.}

\item{lwd.grid}{The line width of the population grid. Defaults to 1.}

\item{add.demes}{A logical value indicating whether to add the observed demes or not.}

\item{col.demes}{The color of the demes. Defaults to 'black'.}

\item{pch.demes}{The symbol, specified as an integer, or the character to be used for plotting the demes. Defaults to 19.}

\item{min.cex.demes,max.cex.demes}{The minimum and the maximum size of the deme symbol/character. Defaults to 1 and 3, respectively. If max.cex.demes > min.cex.demes, then demes with more samples also have bigger size: the deme with the fewest samples has size 'min.cex.demes' and the deme with the most samples has size 'max.cex.demes'.}

\item{add.outline}{A logical value indicating whether to add the habitat outline or not.}

\item{col.outline}{The color of the habitat outline. Defaults to 'white'.}

\item{lwd.outline}{The line width of the habitat outline. Defaults to 2.}

\item{projection.in,projection.out}{The input and the output cartographic projections, specified as a PROJ.4 string. Require the 'rgdal' package.}

\item{add.map}{A logical value indicating whether to add a high-resolution geographic map. Requires the 'rworldmap' and 'rworldxtra' packages. It also requires that 'projection.in' is specified.}

\item{col.map}{The color of the geographic map. Default is 'gray60'.}

\item{lwd.map}{The line width of the geographic map. Defaults to 2.}

\item{eems.colors}{A list of colors to use as the EEMS color scheme, ordered from low to high. Only a divergent palette makes sense. Defaults to a DarkOrange to Blue divergent palette with six orange shades, white and six blue shades.}
}
\description{
Given a list of EEMS output directories, this function generates five figures to visualize EEMS results.
}
\examples{
## Use the provided example or supply the path to your own EEMS run.
eems.results.to.plot = paste(path.package("rEEMSplots"),"/extdata/EEMS-example",sep="")
name.figures.to.save = "EEMS-example-rEEMSplots"

## Produce the five EEMS figures, with default values for all optional parameters.
eems.plots(mcmcpath = eems.results.to.plot,
           plotpath = paste(name.figures.to.save,"-default",sep=""),
           longlat = TRUE)

## Flip the x and y axis, i.e., assume that the x coordinate is the latitude
## and the y coordinate is the longitude.
eems.plots(mcmcpath = eems.results.to.plot,
           plotpath = paste(name.figures.to.save,"-axes-flipped",sep=""),
           longlat = FALSE)

## Generate EEMS figures as png files with height 9 inches, width 8 inches
## and resolution 600 dots per inch.
eems.plots(mcmcpath = eems.results.to.plot,
           plotpath = paste(name.figures.to.save,"-demes-and-edges",sep=""),
           longlat = TRUE,
           plot.height = 9,
           plot.width = 8,
           res = 600,
           out.png = TRUE)

## Generate EEMS figures as pdf files with height 9 inches and width 8 inches.
## The resolution option, res, will be ignored.
eems.plots(mcmcpath = eems.results.to.plot,
           plotpath = paste(name.figures.to.save,"-output-PDFs",sep=""),
           longlat = TRUE,
           plot.height = 9,
           plot.width = 8,
           res = 600,
           out.png = FALSE)

## Choose somewhat impractical colors for the outline, the grid and the demes.
eems.plots(mcmcpath = eems.results.to.plot,
           plotpath = paste(name.figures.to.save,"-demes-and-edges",sep=""),
           longlat = TRUE,
           add.grid = TRUE,
           col.grid = "gray90",
           lwd.grid = 2,
           add.outline = TRUE,
           col.outline = "blue",
           lwd.outline = 5,
           add.demes = TRUE,
           col.demes = "red",
           pch.demes = 5,
           min.cex.demes = 0.5,
           max.cex.demes = 1.5)

library(rgdal)

## Produce contour plots in the Mercator projection (used by Google Maps)
eems.plots(mcmcpath = eems.results.to.plot,
           plotpath = paste(name.figures.to.save,"-cartographic-projections",sep=""),
           longlat = TRUE,
           projection.in = "+proj=longlat +datum=WGS84",
           projection.out = "+proj=merc +datum=WGS84")

library(rworldmap)
library(rworldxtra)

## Add a high-resolution geographic map
eems.plots(mcmcpath = eems.results.to.plot,
           plotpath = paste(name.figures.to.save,"-geographic-map",sep=""),
           longlat = TRUE,
           projection.in = "+proj=longlat +datum=WGS84",
           projection.out = "+proj=merc +datum=WGS84",
           add.map = TRUE,
           col.map = "black",
           lwd.map = 5)

library(RColorBrewer)

## Use a divergent Red to Blue color scheme from the 'RColorBrewer' package
## instead of the default DarkOrange to Blue color scheme.
## In 'RColorBrewer', the palette runs from Dark Red to Dark Blue but it is
## more intuitive to use shades of blue for lower-than-average rates and
## shades of red for higher-than-average rates, so the "RdBu" theme is
## reversed to a "BuRd" theme, with the 'rev' function.
eems.plots(mcmcpath = eems.results.to.plot,
           plotpath = paste(name.figures.to.save,"-new-eems-colors",sep=""),
           longlat = TRUE,
           projection.in = "+proj=longlat +datum=WGS84",
           projection.out = "+proj=merc +datum=WGS84",
           eems.colors = rev(brewer.pal(11,"RdBu")))
}
\keyword{rEEMSplots}

