% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rEEMSplots.R
\name{eems.resid.heatmap}
\alias{eems.resid.heatmap}
\title{A function to plot a heatmap of the residual pairwise dissimilarities, abs(observed - fitted)}
\usage{
eems.resid.heatmap(      datapath,
                         mcmcpath,
                         plotpath,

                         ## Properties of the figure to generate
                         plot.width = 7,
                         plot.height = 7,
                         out.png = TRUE,
                         res = 600,

                         ## (Divergent) color palette
                         heatmap.cols = NULL,
                         heatmap.colscale = NULL)
}
\arguments{
\item{datapath}{The full path and the file name of the input Diffs matrix, which is not copied by \code{runeems} to the output directory.}

\item{mcmcpath}{An EEMS output directory. If mcmcpath is a vector of directories, then only the first (existing) directory is used.}

\item{plotpath}{The full path and the file name for the graphics to be generated.}

\item{plot.width,plot.height}{The width and height of the graphics region for the two rate contour plots, in inches. The default values are both 7.}

\item{out.png}{A logical value indicating whether to generate output graphics as PNGs (the default) or PDFs.}

\item{res}{Resolution, in dots per inch; used only if \code{out.png} is set to TRUE. The default is 600.}

\item{heatmap.cols}{The heatmap color palette as a vector of colors, ordered from low to high. Defaults to the "Reds" divergent palette in the RColorBrewer package.}
}
\description{
Given a vector of EEMS output directories, this function generates a heatmap of the n-by-n matrix of residual dissimilarities between pairs of individuals. It also saves the matrix of residuals to a file called \code{plotpath}-eems-resid.RData. In both the residual matrix and the corresponding heatmap, individuals are ordered according to the input \code{datapath}.coord and \code{datapath}.diffs. Applies only to the case of SNP data (when a Diffs matrix is computed).
}
\examples{
## Use the provided example or supply the path to your own EEMS run
eems.dataset.to.plot = paste(path.package("rEEMSplots"),"/extdata/EEMS-barrier",sep="")
eems.results.to.plot = paste(path.package("rEEMSplots"),"/extdata/EEMS-barrier",sep="")
name.figures.to.save = "EEMS-barrier"

eems.resid.heatmap(datapath = eems.dataset.to.plot,
                   mcmcpath = eems.results.to.plot,
                   plotpath = name.figures.to.save,
                   heatmap.cols = c("gray99","red"))
}

