% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recenter.longitudes.R
\name{recenter_by}
\alias{recenter_by}
\title{Shift spatial object longitudinally}
\usage{
recenter_by(obj, degree)
}
\arguments{
\item{obj}{A SpatialPolygons or a SpatialPolygonsDataFrame object.}

\item{degree}{Number of degrees (positive or negative) to shift.}
}
\description{
The function `sp::recenter` shifts coordinates for a Pacific view by shifting longitudes
from [-180, 180] to [0, 360]. I modified `sp::recenter` to shift longitudes by any degree.
}
