% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour.with.holes.R
\name{remove_demes_outside_habitat}
\alias{remove_demes_outside_habitat}
\title{Remove demes that lie outside of the habitat}
\usage{
remove_demes_outside_habitat(habitat, mcmcpath, gridpath)
}
\arguments{
\item{habitat}{A SpatialPolygons habitat (possibly with holes).}

\item{mcmcpath}{An EEMS output directory which contrains at a minimum the files
demes.txt, edges.txt, ipmap.txt.}

\item{gridpath}{Output filename.}
}
\description{
Remove demes that lie outside of the habitat
}
