% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rEEMSplots.R
\name{voronoi.plots}
\alias{voronoi.plots}
\title{A function to plot Voronoi diagrams of migration and diversity rates}
\usage{
voronoi.plots(mcmcpath, plotpath, longlat, plot.width = 7, plot.height = 7,
  out.png = TRUE, res = 600, eems.colors = NULL)
}
\arguments{
\item{mcmcpath}{An EEMS output directory. If mcmcpath is a list of directories, then only the first (existing) directory in the list is used.}

\item{plotpath}{The name of the graphics files to generate. The output is a series of posterior Voronoi diagrams, one series for the migration rates and another for the diversity rates, with two figures for each saved MCMC iteration (after the burn-in and thinning). Warning: Potentially generates a largen number of figures.}

\item{longlat}{A logical value indicating whether the coordinates are given as pairs (longitude, latitude) or (latitude, longitude).}

\item{plot.width,plot.height}{The width and height of the graphics region for the two rate contour plots, in inches. The default values are both 7.}

\item{out.png}{A logical value indicating whether to generate png files or pdf graphics files.}

\item{res}{Resolution, in dots per inch; used only if out.png is set to TRUE. The default is 600.}

\item{eems.colors}{A list of colors to use as the EEMS color scheme, ordered from low to high. Only a divergent palette makes sense. Defaults to a DarkOrange to Blue divergent palette with six orange shades, white and six blue shades.}
}
\description{
A function to plot Voronoi diagrams of migration and diversity rates
}
\examples{
## Use the provided example or supply the path to your own EEMS run.
eems.results.to.plot = paste(path.package("rEEMSplots"),"/extdata/EEMS-example",sep="")
name.figures.to.save= "EEMS-example-rEEMSplots"

library(deldir)

voronoi.plots(mcmcpath = eems.results.to.plot,
              plotpath = paste(name.figures.to.save,"-voronoi-diagrams",sep=""),
              longlat = TRUE)
}
\keyword{rEEMSplots}

