% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEEMSplots.R
\name{eems.plots}
\alias{eems.plots}
\title{A function to plot effective migration and diversity surfaces from EEMS output}
\usage{
eems.plots(            mcmcpath, 
                       plotpath, 
                       longlat, 
                       
                       ## Properties of the figures to generate
                       plot.width = 7, 
                       plot.height = 7, 
                       out.png = TRUE, 
                       res = 600, 
                       xpd = TRUE, 
                       
                       ## Properties of the population grid
                       add.grid = FALSE, 
                       col.grid = "gray80", 
                       lwd.grid = 1, 
                       
                       ## Properties of the observed demes
                       add.demes = FALSE, 
                       col.demes = "black", 
                       pch.demes = 19, 
                       min.cex.demes = 1, 
                       max.cex.demes = 3, 
                       
                       ## Properties of the habitat outline
                       add.outline = FALSE, 
                       col.outline = "gray90", 
                       lwd.outline = 2, 
                       
                       ## Properties of the cartographic projection
                       projection.in = NULL, 
                       projection.out = NULL, 
                       
                       ## Properties of the geographic map
                       add.map = FALSE, 
                       col.map = "gray60", 
                       lwd.map = 2, 
                       
                       ## Color palette
                       eems.colors = NULL, 
                       ## Breaks for P(m > 0 | diffs) and P(m < 0 | diffs)
                       prob.levels = c(0.9, 0.95),
                       
                       ## Properties of the color key
                       add.colbar = TRUE, 
                       m.colscale = NULL, 
                       q.colscale = NULL, 
                       
                       ## Remove demes with a single observation
                       remove.singletons = TRUE, 
                       
                       ## Add the line y = x and/or the R squared coefficient to scatter plots
                       add.abline = FALSE,
                       add.r.squared = FALSE,
                       
                       ## Highlight some points (not available yet)
                       highlight.demes = NULL, 
                       
                       ## Extra options
                       add.title = TRUE, 
                       m.plot.xy = NULL, 
                       q.plot.xy = NULL)
}
\arguments{
\item{mcmcpath}{A vector of EEMS output directories, for the same dataset. Warning: There is minimal checking that the given  directories are for the same dataset.}

\item{plotpath}{The full path and the file name for the graphics to be generated.}

\item{longlat}{A logical value indicating whether the coordinates are given as pairs (longitude, latitude) or (latitude, longitude).}

\item{plot.width, }{plot.height The width and height of the graphics region for the two rate contour plots, in inches. The default values are both 7.}

\item{out.png}{A logical value indicating whether to generate output graphics as PNGs (the default) or PDFs.}

\item{res}{Resolution, in dots per inch; used only if \code{out.png} is set to TRUE. The default is 600.}

\item{xpd}{A logical value indicating whether to clip plotting to the figure region (\code{xpd} = TRUE, which is the default) or clip plotting to the plot region (\code{xpd} = FALSE).}

\item{add.grid}{A logical value indicating whether to add the population grid or not.}

\item{col.grid}{The color of the population grid. Defaults to \code{gray80}.}

\item{lwd.grid}{The line width of the population grid. Defaults to 1.}

\item{add.demes}{A logical value indicating whether to add the observed demes or not.}

\item{col.demes}{The color of the demes. Defaults to \code{black}.}

\item{pch.demes}{The symbol, specified as an integer, or the character to be used for plotting the demes. Defaults to 19.}

\item{min.cex.demes, }{max.cex.demes The minimum and the maximum size of the deme symbol/character. Defaults to 1 and 3, respectively. If \code{max.cex.demes} > \code{min.cex.demes}, then demes with more samples also have bigger size: the deme with the fewest samples has size \code{min.cex.demes} and the deme with the most samples has size \code{max.cex.demes}.}

\item{add.outline}{A logical value indicating whether to add the habitat outline or not.}

\item{col.outline}{The color of the habitat outline. Defaults to \code{white}.}

\item{lwd.outline}{The line width of the habitat outline. Defaults to 2.}

\item{projection.in, }{projection.out The input and the output cartographic projections, each specified as a PROJ.4 string. Requires the \code{rgdal} package.}

\item{add.map}{A logical value indicating whether to add a high-resolution geographic map. Requires the \code{rworldmap} and \code{rworldxtra} packages. It also requires that \code{projection.in} is specified.}

\item{col.map}{The color of the geographic map. Default is \code{gray60}.}

\item{lwd.map}{The line width of the geographic map. Defaults to 2.}

\item{eems.colors}{The EEMS color scheme as a vector of colors, ordered from low to high. Defaults to a DarkOrange to Blue divergent palette with six orange shades, white in the middle, six blue shades. Acknowledgement: The default color scheme is adapted from the \code{dichromat} package.}

\item{prob.levels}{A vector of probabilities for plotting the posterior probability contours of \code{P(m > 0 | diffs)} and \code{P(m < 0 | diffs)}. Defaults to \code{c(0.9, 0.95)}.}

\item{add.colbar}{A logical value indicating whether to add the color bar (the key that shows how colors map to rates) to the right of the plot. Defaults to TRUE.}

\item{m.colscale, }{q.colscale A fixed range for log10-transformed migration and diversity rates, respectively. If the estimated rates fall outside the specified range, then the color scale is ignored. By default, no range is specified for either type of rates.}

\item{remove.singletons}{Remove demes with a single observation from the diagnostic scatter plots. Defaults to TRUE.}

\item{add.abline}{Add the line \code{y = x} to the diagnostic scatter plots of observed vs fitted genetic dissimilarities.}

\item{add.r.squared}{Add the R squared coefficient to the diagnostic scatter plots of observed vs fitted genetic dissimilarities.}

\item{add.title}{A logical value indicating whether to add the main title in the contour plots. Defaults to TRUE.}
}
\description{
Given a vector of EEMS output directories, this function generates several figures to visualize EEMS results. It is a good idea to examine all these figures, which is why they are generated by default. 
\itemize{
 \item \code{plotpath-mrates01}: effective migration surface. This contour plot visualizes the estimated effective migration rates \code{m}, on the log10 scale after mean centering.
 \item \code{plotpath-mrates02}: posterior probabilities \code{P(m > 0 | diffs)} and \code{P(m < 0 | diffs)} for each location in the habitat. Since migration rates are visualized on the log10 scale after mean centering, 0 corresponds to the overall mean migration rate. This contour plot emphasizes regions with effective migration that is significantly higher/lower than the overall average.
 \item \code{plotpath-qrates01}: effective diversity surface. This contour plot visualizes the estimated effective diversity rates \code{q}, on the log10 scale after mean centering.
 \item \code{plotpath-qrates02}: posterior probabilities \code{P(q > 0 | diffs)} and \code{P(q < 0 | diffs)}. Similar to \code{plotpath-mrates02} but applied to the effective diversity rates.
 \item \code{plotpath-rdist01}: scatter plot of the observed vs the fitted between-deme component of genetic dissimilarity, where one point represents a pair of sampled demes.
 \item \code{plotpath-rdist01}: scatter plot of the observed vs the fitted within-deme component of genetic dissimilarity, where one point represents a sampled deme.
 \item \code{plotpath-rdist03}: scatter plot of observed genetic dissimilarities between demes vs observed geographic distances between demes.
 \item \code{plotpath-pilogl01}: posterior probability trace
}
The \code{mrates} and \code{qrates} figures visualize (properties of) the effective migration and diversity rates across the habitat. The other figures can help to check that the MCMC sampler has converged (the trace plot \code{pilogl}) and that the EEMS model fits the data well (the scatter plots of genetic dissimilarities \code{rdist}).
}
\details{
The function \code{eems.plots} will work given the results from a single EEMS run (one directory in \code{mcmcpath}) but it is better to run EEMS several times, randomly initializing the MCMC chain each time. In other words, simulate several realizations of the Markov chain and let each realization start from a different state in the parameter space (by using a different random seed).

Detail about the within-deme and between-deme components of genetic dissimilarity: Let \code{D(a,b)} be the dissimilarity between one individual from deme \code{a} and another individual from deme \code{b}. Then the within-deme component for \code{a} and \code{b} is simply \code{D(a,a)} and \code{D(b, b)}, respectively. The between-deme component is \code{D(a,b) - [D(a,a) + D(b,b)] / 2} and it represents dissimilarity that is due to the spatial structure of the population and is not a consequence of the local diversity in the two demes.
}
\examples{

## Use the provided example or supply the path to your own EEMS run.
eems.results.to.plot = paste0(path.package("rEEMSplots"), "/extdata/EEMS-example")
name.figures.to.save = path.expand("~/EEMS-example-rEEMSplots")

## Produce the five EEMS figures, with default values for all optional parameters.
eems.plots(mcmcpath = eems.results.to.plot, 
           plotpath = paste0(name.figures.to.save, "-default"), 
           longlat = TRUE)

## Flip the x and y axis, i.e., assume that the x coordinate is the latitude
## and the y coordinate is the longitude.
eems.plots(mcmcpath = eems.results.to.plot, 
           plotpath = paste0(name.figures.to.save, "-axes-flipped"), 
           longlat = FALSE)

## Generate PNG figures with height 9 inches, width 8 inches
## and resolution 600 dots per inch.
eems.plots(mcmcpath = eems.results.to.plot, 
           plotpath = paste0(name.figures.to.save, "-demes-and-edges"), 
           longlat = TRUE, 
           plot.height = 8, 
           plot.width = 7, 
           res = 600, 
           out.png = TRUE)

## Generate PDF figures with height 9 inches and width 8 inches.
## The resolution option, res, is ignored.
eems.plots(mcmcpath = eems.results.to.plot, 
           plotpath = paste0(name.figures.to.save, "-output-PDFs"), 
           longlat = TRUE, 
           plot.height = 8, 
           plot.width = 7, 
           res = 600, 
           out.png = FALSE)

## Choose somewhat impractical colors and shapes for the outline, the grid and the demes.
eems.plots(mcmcpath = eems.results.to.plot, 
           plotpath = paste0(name.figures.to.save, "-demes-and-edges"), 
           longlat = TRUE, 
           add.grid = TRUE, 
           col.grid = "gray90", 
           lwd.grid = 2, 
           add.outline = TRUE, 
           col.outline = "blue", 
           lwd.outline = 5, 
           add.demes = TRUE, 
           col.demes = "red", 
           pch.demes = 5, 
           min.cex.demes = 0.5, 
           max.cex.demes = 1.5)

library("rgdal")

## Produce contour plots in the Mercator projection (used by Google Maps)
eems.plots(mcmcpath = eems.results.to.plot, 
           plotpath = paste0(name.figures.to.save, "-cartographic-projections"), 
           longlat = TRUE, 
           projection.in = "+proj=longlat +datum=WGS84", 
           projection.out = "+proj=merc +datum=WGS84")

library("rworldmap")
library("rworldxtra")

## Add a high-resolution geographic map
eems.plots(mcmcpath = eems.results.to.plot, 
           plotpath = paste0(name.figures.to.save, "-geographic-map"), 
           longlat = TRUE, 
           projection.in = "+proj=longlat +datum=WGS84", 
           projection.out = "+proj=merc +datum=WGS84", 
           add.map = TRUE, 
           col.map = "black", 
           lwd.map = 5)

library("RColorBrewer")

## Use a divergent Red to Blue color scheme from the RColorBrewer package
## instead of the default DarkOrange to Blue color scheme.
eems.plots(mcmcpath = eems.results.to.plot, 
           plotpath = paste0(name.figures.to.save, "-new-eems-colors"), 
           longlat = TRUE, 
           projection.in = "+proj=longlat +datum=WGS84", 
           projection.out = "+proj=merc +datum=WGS84", 
           eems.colors = brewer.pal(11, "RdBu"))

## Specify the color scales for the migration and diversity rates 
eems.plots(mcmcpath = eems.results.to.plot, 
           plotpath = paste0(name.figures.to.save, "-fix-colscales"), 
           longlat = TRUE, 
           add.outline = TRUE, 
           col.outline = "gray", 
           m.colscale = c(-3, 3), 
           q.colscale = c(-0.3, +0.3))

## In the diagnostic scatter plots, highlight demes in the South in red
## All demes in the population grid, as (longitude, latitude) pairs
demes = read.table(paste0(eems.results.to.plot, "/demes.txt"))
demes.long = demes[, 1]
demes.lat = demes[, 2]
## The demes in the South, defined as the region below the 10th parallel south
south = demes[demes.lat < -10, ]
highlight.demes = data.frame(
    x = south[, 1], 
    y = south[, 2], 
    col = "red", 
    stringsAsFactors = FALSE)
eems.plots(mcmcpath = eems.results.to.plot, 
           plotpath = paste0(name.figures.to.save, "-highlight-demes"), 
           longlat = TRUE, 
           highlight.demes = highlight.demes)
           
}
\references{
Light A and Bartlein PJ (2004). The End of the Rainbow? Color Schemes for Improved Data Graphics. EOS Transactions of the American Geophysical Union, 85(40), 385.
}
\seealso{
\code{\link{eems.voronoi.samples}, \link{eems.posterior.draws}, \link{eems.resid.heatmap}, \link{eems.population.grid}}
}

