% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rEEMSplots.R
\name{eems.plots}
\alias{eems.plots}
\title{A function to plot effective migration and diversity surfaces from EEMS output}
\usage{
eems.plots(mcmcpath, plotpath, longlat, plot.width = 7, plot.height = 7,
  out.png = TRUE, res = 600, xpd = TRUE, add.grid = FALSE,
  col.grid = "gray80", lwd.grid = 1, add.demes = FALSE,
  col.demes = "black", pch.demes = 19, min.cex.demes = 1,
  max.cex.demes = 3, add.outline = FALSE, col.outline = "white",
  lwd.outline = 2, projection.in = NULL, projection.out = NULL,
  add.map = FALSE, col.map = "gray60", lwd.map = 2, eems.colors = NULL)
}
\arguments{
\item{mcmcpath}{A vector of EEMS output directories, for the same dataset. Warning: There is minimal checking that the given  directories are for the same dataset.}

\item{plotpath}{The full path and the file name for the graphics to be generated.}

\item{longlat}{A logical value indicating whether the coordinates are given as pairs (longitude, latitude) or (latitude, longitude).}

\item{plot.width,plot.height}{The width and height of the graphics region for the two rate contour plots, in inches. The default values are both 7.}

\item{out.png}{A logical value indicating whether to generate output graphics as PNGs (the default) or PDFs.}

\item{res}{Resolution, in dots per inch; used only if \code{out.png} is set to TRUE. The default is 600.}

\item{xpd}{A logical value indicating whether to clip plotting to the figure region (\code{xpd} = TRUE, which is the default) or clip plotting to the plot region (\code{xpd} = FALSE).}

\item{add.grid}{A logical value indicating whether to add the population grid or not.}

\item{col.grid}{The color of the population grid. Defaults to \code{gray80}.}

\item{lwd.grid}{The line width of the population grid. Defaults to 1.}

\item{add.demes}{A logical value indicating whether to add the observed demes or not.}

\item{col.demes}{The color of the demes. Defaults to \code{black}.}

\item{pch.demes}{The symbol, specified as an integer, or the character to be used for plotting the demes. Defaults to 19.}

\item{min.cex.demes,max.cex.demes}{The minimum and the maximum size of the deme symbol/character. Defaults to 1 and 3, respectively. If \code{max.cex.demes} > \code{min.cex.demes}, then demes with more samples also have bigger size: the deme with the fewest samples has size \code{min.cex.demes} and the deme with the most samples has size \code{max.cex.demes}.}

\item{add.outline}{A logical value indicating whether to add the habitat outline or not.}

\item{col.outline}{The color of the habitat outline. Defaults to \code{white}.}

\item{lwd.outline}{The line width of the habitat outline. Defaults to 2.}

\item{projection.in,projection.out}{The input and the output cartographic projections, each specified as a PROJ.4 string. Requires the \code{rgdal} package.}

\item{add.map}{A logical value indicating whether to add a high-resolution geographic map. Requires the \code{rworldmap} and \code{rworldxtra} packages. It also requires that \code{projection.in} is specified.}

\item{col.map}{The color of the geographic map. Default is \code{gray60}.}

\item{lwd.map}{The line width of the geographic map. Defaults to 2.}

\item{eems.colorsh}{The EEMS color scheme as a vector of colors, ordered from low to high. Defaults to a DarkOrange to Blue divergent palette with six orange shades, white in the middle, six blue shades. Acknowledgement: The default color scheme is adapted from the \code{dichromat} package.}
}
\description{
Given a vector of EEMS output directories, this function generates five figures to visualize EEMS results:
\itemize{
 \item \code{plotpath}-mrates01 (effective migration surface)
 \item \code{plotpath}-qrates01 (effective diversity surface)
 \item \code{plotpath}-rdist01 (between-demes component of genetic dissimilarity)
 \item \code{plotpath}-rdist02 (within-demes component of genetic dissimilarity)
 \item \code{plotpath}-pilogl01 (posterior probability trace)
}
The three latter figures are helpful in checking that the MCMC sampler has converged (the trace plot) and that the EEMS model fits the data well (the scatter plots of genetic dissimilarities). \code{eems.plots} will work with a single EEMS output directory but it is better to run EEMS several times, randomly initializing the MCMC chain each time. In other words, it is a good idea to simulate several realizations of the Markov chain, each realization starting with a different value of the EEMS parameters.
}
\examples{
## Use the provided example or supply the path to your own EEMS run.
eems.results.to.plot = paste(path.package("rEEMSplots"),"/extdata/EEMS-example",sep="")
name.figures.to.save = "EEMS-example-rEEMSplots"

## Produce the five EEMS figures, with default values for all optional parameters.
eems.plots(mcmcpath = eems.results.to.plot,
           plotpath = paste(name.figures.to.save,"-default",sep=""),
           longlat = TRUE)

## Flip the x and y axis, i.e., assume that the x coordinate is the latitude
## and the y coordinate is the longitude.
eems.plots(mcmcpath = eems.results.to.plot,
           plotpath = paste(name.figures.to.save,"-axes-flipped",sep=""),
           longlat = FALSE)

## Generate PNG figures with height 9 inches, width 8 inches
## and resolution 600 dots per inch.
eems.plots(mcmcpath = eems.results.to.plot,
           plotpath = paste(name.figures.to.save,"-demes-and-edges",sep=""),
           longlat = TRUE,
           plot.height = 9,
           plot.width = 8,
           res = 600,
           out.png = TRUE)

## Generate PDF figures with height 9 inches and width 8 inches.
## The resolution option, res, is ignored.
eems.plots(mcmcpath = eems.results.to.plot,
           plotpath = paste(name.figures.to.save,"-output-PDFs",sep=""),
           longlat = TRUE,
           plot.height = 9,
           plot.width = 8,
           res = 600,
           out.png = FALSE)

## Choose somewhat impractical colors and shapes for the outline, the grid and the demes.
eems.plots(mcmcpath = eems.results.to.plot,
           plotpath = paste(name.figures.to.save,"-demes-and-edges",sep=""),
           longlat = TRUE,
           add.grid = TRUE,
           col.grid = "gray90",
           lwd.grid = 2,
           add.outline = TRUE,
           col.outline = "blue",
           lwd.outline = 5,
           add.demes = TRUE,
           col.demes = "red",
           pch.demes = 5,
           min.cex.demes = 0.5,
           max.cex.demes = 1.5)

library(rgdal)

## Produce contour plots in the Mercator projection (used by Google Maps)
eems.plots(mcmcpath = eems.results.to.plot,
           plotpath = paste(name.figures.to.save,"-cartographic-projections",sep=""),
           longlat = TRUE,
           projection.in = "+proj=longlat +datum=WGS84",
           projection.out = "+proj=merc +datum=WGS84")

library(rworldmap)
library(rworldxtra)

## Add a high-resolution geographic map
eems.plots(mcmcpath = eems.results.to.plot,
           plotpath = paste(name.figures.to.save,"-geographic-map",sep=""),
           longlat = TRUE,
           projection.in = "+proj=longlat +datum=WGS84",
           projection.out = "+proj=merc +datum=WGS84",
           add.map = TRUE,
           col.map = "black",
           lwd.map = 5)

library(RColorBrewer)

## Use a divergent Red to Blue color scheme from the RColorBrewer package
## instead of the default DarkOrange to Blue color scheme.
eems.plots(mcmcpath = eems.results.to.plot,
           plotpath = paste(name.figures.to.save,"-new-eems-colors",sep=""),
           longlat = TRUE,
           projection.in = "+proj=longlat +datum=WGS84",
           projection.out = "+proj=merc +datum=WGS84",
           eems.colors = brewer.pal(11,"RdBu"))
}
\references{
Light A and Bartlein PJ (2004). The End of the Rainbow? Color Schemes for Improved Data Graphics. EOS Transactions of the American Geophysical Union, 85(40), 385.
}

