% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eems.resid.heatmap.R
\name{eems.resid.heatmap}
\alias{eems.resid.heatmap}
\title{A function to plot a heatmap of the residual pairwise dissimilarities, abs(observed - fitted)}
\usage{
eems.resid.heatmap(
  datapath,
  mcmcpath,
  plotpath,
  plot.width = 7,
  plot.height = 7,
  out.png = TRUE,
  res = 600,
  heatmap.cols = NULL,
  heatmap.colscale = NULL
)
}
\arguments{
\item{datapath}{The full path and the file name of the input Diffs matrix, which is not copied
by \code{runeems} to the output directory.}

\item{mcmcpath}{A vector of EEMS output directories, for the same dataset. Warning: There is
minimal checking that the given  directories are for the same dataset.}

\item{plotpath}{The full path and the file name for the graphics to be generated.}

\item{plot.width}{The width of the graphics region for the two rate contour plots, in inches.
The default value is 7.}

\item{plot.height}{The height of the graphics region, in inches. The default value is 7.}

\item{out.png}{A logical value indicating whether to generate output graphics as PNGs
(the default) or PDFs.}

\item{res}{Resolution, in dots per inch; used only if \code{out.png} is set to TRUE.
The default is 600.}

\item{heatmap.cols}{The heatmap color palette as a vector of colors, ordered from low to high.
Defaults to the "Reds" divergent palette in the RColorBrewer package.}

\item{heatmap.colscale}{A fixed range for the heatmap colors. The default is NULL, so the color
space is the observed range of the residuals.}
}
\description{
Given a set of EEMS output directories, this function generates a heatmap of the n-by-n matrix
of residual dissimilarities between pairs of individuals. The residuals are the differences
between the observed and the fitted dissimilarities.
The function also saves the residual matrix to a file called \code{plotpath-eems-resid.RData}.
In both the residual matrix and the corresponding heat map, individuals are are in the same
order as in the input files \code{datapath.coord} and \code{datapath.diffs}. Applicable only
in the case of SNP data when the observed dissimilarity matrix is computed explicitly.
}
\examples{
# Use the provided example or supply the path to your own EEMS run
extdata_path <- system.file("extdata", package = "rEEMSplots")
eems_dataset <- file.path(extdata_path, "EEMS-barrier")
eems_results <- file.path(extdata_path, "EEMS-barrier")
name_figures <- file.path(path.expand("~"), "EEMS-barrier")

eems.resid.heatmap(
  datapath = eems_dataset,
  mcmcpath = eems_results,
  plotpath = name_figures,
  heatmap.cols = c("gray99", "red")
)
}
\seealso{
\code{\link{eems.plots}, \link{eems.voronoi.samples}, \link{eems.posterior.draws},
\link{eems.population.grid}}
}
