% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rEEMSplots.R
\name{eems.voronoi}
\alias{eems.voronoi}
\title{A function to plot Voronoi diagrams of effective migration and diversity rates}
\usage{
eems.voronoi(mcmcpath, plotpath, longlat, plot.width = 7, plot.height = 7,
  out.png = TRUE, res = 600, eems.colors = NULL)
}
\arguments{
\item{mcmcpath}{An EEMS output directory. If mcmcpath is a vector of directories, then only the first (existing) directory is used.}

\item{plotpath}{The full path and the file name for the graphics to be generated.}

\item{longlat}{A logical value indicating whether the coordinates are given as pairs (longitude, latitude) or (latitude, longitude).}

\item{plot.width,plot.height}{The width and height of the graphics region for the two rate contour plots, in inches. The default values are both 7.}

\item{out.png}{A logical value indicating whether to generate output graphics as PNGs (the default) or PDFs.}

\item{res}{Resolution, in dots per inch; used only if \code{out.png} is set to TRUE. The default is 600.}

\item{eems.colorsh}{The EEMS color scheme as a vector of colors, ordered from low to high. Defaults to a DarkOrange to Blue divergent palette with six orange shades, white in the middle, six blue shades. Acknowledgement: The default color scheme is adapted from the \code{dichromat} package.}
}
\description{
Given one EEMS output directory, this function produces a series of Voronoi diagrams drawn from the posterior distribution of the rate parameters. There is one series of tessellations for the effective migration rates and another series for the effective diversity rates. Both series contain one figure for each saved MCMC iteration, after burn-in and thinning. Warning: \code{eems.voronoi} potentially generates a large number of figures.
}
\examples{
## Use the provided example or supply the path to your own EEMS run.
eems.results.to.plot = paste(path.package("rEEMSplots"),"/extdata/EEMS-example",sep="")
name.figures.to.save= "EEMS-example-rEEMSplots"

library(deldir)

## Plot a series of Voronoi diagrams for the EEMS model parameters:
## the effective migration rates (m) and the effective diversity rates (q).
eems.voronoi(mcmcpath = eems.results.to.plot,
            plotpath = paste(name.figures.to.save,"-voronoi-diagrams",sep=""),
            longlat = TRUE)
}
\references{
Light A and Bartlein PJ (2004). The End of the Rainbow? Color Schemes for Improved Data Graphics. EOS Transactions of the American Geophysical Union, 85(40), 385.
}

